Added regex pattern matching method
I'm assuming this is how we bring it in - however do you think this is the best way to handle the additional methods (there are many more, so arguably you'd have to include them all)? Admittedly `pattern` is a very useful one though, as it's kind of a 'do anything' method, so if you're going to include any additional method, that's probably the one to include.

I have a feeling what would be more useful is if you could activate additional methods as options when using the meteor package - however that's *way* outside my comfort zone, so just a suggestion!

If there's a way to achieve the above in my own app without hacking the package I'd also be fine with that - however I had a quick look at manually including files from packages and it didn't look straightforward.

(Oh, and incidentally, I couldn't see a `src/delegate.js` file in that location (line 13) - maybe it's moved? I didn't want to remove it myself in case I was missing something.)